Creates CreditCardFormsController to generate credit card GPOs
The correct way to add a specialized view like the credit card GPOs is
not, as I once believed, to add methods to the controller it almost
makes sense to be in. Better is to create a new controller to handle
specifically that use case. Since displaying a credit card form is
something different from displaying donation information, I made a
CreditCardFormsController.

It's simple: only `:show` and `:new`. The trick is in the routes.
It's best to use 'resources' as much as possible in routes, since it
generates url helpers for you, which is super convenient. Check out what
I did in `/config/routes.rb`. The big trick is the [singular resource](http://guides.rubyonrails.org/routing.html#singular-resources)
to get `/donations/:donation_id/credit_card_form`